Fix bug in BinFileReader.bytePair2Char
The bytePair2Char was not converting the bytes to ints with the unsigned
value of the byte first before
combining each byte pair into a char.  This caused issues with negative
bytes.  By ANDing each byte with 0xFFFF this converted the byte to an
int with the unsigned value of the byte.