Fix Cell to hide Bytes and return Ints
I previously had a bug converting Byte to Int.  By always doing the
conversion inside the class and forcing clients to use the public API
that always returns the correct Int, the bug should be prevented in the
future for reoccurring.

The previous bug:
Tile index is a byte; which .toInt does not return 0-255;
(ex: 255 -> -1)

The fix:
Change class to store byte internally, but give api that
returns proper int (same for bg & fg colors)